Changed return types on currentWeight function

+++ libstreetmap/src/m4Helpers.cpp
@@ -12,8 +12,8 @@ double percentFull(double maxWeight, double currentWeight){
     return (currentWeight/maxWeight * 100); 
 }
 
-double currentWeight(const std::vector<DeliveryInfo>& deliveries, std::vector<int> deliveryIndices){
-    double weight = 0; 
+float currentWeight(const std::vector<DeliveryInfo>& deliveries, std::vector<int> deliveryIndices){
+    float weight = 0; 
     for (int i = 0; i < deliveryIndices.size(); i++){
         weight += deliveries[deliveryIndices[i]].itemWeight; 
     }
